Cleaned code, Separate file for Populating Functions, used funcs for adding
