+++ CCC151-Group--BH-Management-System/Py_Files/ADD/AddTenantDialog.py
@@ -1,7 +1,7 @@
 from PyQt5.QtWidgets import QDialog, QMessageBox, QCompleter
 from .AddTenant import Ui_Dialog
 import re
-from DATABASE.Functions import Select, Insert, update
+from DATABASE.Functions import Select, Insert, update, Populate
 from DATABASE.DB import DatabaseConnector
 
 class AddTenantDialog(QDialog):
@@ -13,45 +13,13 @@ class AddTenantDialog(QDialog):
         self.select = Select.Select()
         self.insert = Insert.Insert()
         self.update = update.update()
+        self.populate = Populate.Populate()
 
         self.ui.CancelpushButton.clicked.connect(self.reject)
         self.ui.AddpushButton.clicked.connect(self.handle_add_tenant)
 
-        self.populate_sex_combobox()
-        self.populate_room_combobox()
-
-    def populate_sex_combobox(self):
-        self.ui.SexComboBox.clear()
-        self.ui.SexComboBox.addItems(['Male', 'Female'])
-
-    def populate_room_combobox(self):
-        try:
-            # Querying for room numbers
-            rows = self.select.SelectQuery("Room", spec_col=["RoomNumber"]).retData()
-
-            # Debugging: Print the query result
-            # print(f"Rows: {rows}, Columns: {columns}")
-
-            if not rows:
-                QMessageBox.warning(self, "No Data", "No room numbers found.")
-                return
-
-            # Extract room numbers from the query result
-            room_list = [str(row[0]) for row in rows]  # Assuming row[0] contains RoomNumber
-
-            # Populate the combo box with room numbers
-            self.ui.RoomNoComboBox.clear()
-            self.ui.RoomNoComboBox.addItems(room_list)
-
-            # Make combo box searchable
-            self.ui.RoomNoComboBox.setEditable(True)
-            completer = QCompleter(room_list, self.ui.RoomNoComboBox)
-            completer.setCaseSensitivity(False)
-            self.ui.RoomNoComboBox.setCompleter(completer)
-
-        except Exception as err:
-            print(f"Database Error: {err}")
-            QMessageBox.critical(self, "Database Error", f"Failed to load rooms:\n{err}")
+        self.populate.populate_sex_combobox(self.ui.SexComboBox)
+        self.populate.populate_room_combobox(self.ui.RoomNoComboBox)
         
     def handle_add_tenant(self):
         tenant_fname = self.ui.FirstNameLineEdit.text()
@@ -79,18 +47,16 @@ class AddTenantDialog(QDialog):
             return
 
         try:
-            # Create an instance of the Select class
 
             columns = ["MaximumCapacity", "NoOfOccupants", "TenantSex"]
             result = self.select.SelectQuery("Room", spec_col=columns, tag = "RoomNumber", key = tenant_room, limit = 1).retData()
-            print(result)
 
             if result:
                 maximum_capacity, current_occupants, room_tsex = result[0]
                 if room_tsex != tenant_sex and room_tsex != None:
                     QMessageBox.warning(self, "Sex invalid", f"Room {tenant_room} only accepts {room_tsex} tenants.")
                 if current_occupants >= maximum_capacity:
-                    QMessageBox.warning(self, "Room Full", "The room has reached its maximum capacity.")
+                    QMessageBox.warning(self, "Room Full", f"Room {tenant_room} has reached its maximum capacity of {current_occupants}/{maximum_capacity}.")
                     return
                 
             newTen = [
@@ -108,7 +74,7 @@ class AddTenantDialog(QDialog):
 
             self.update.updateTableData("Room", {"NoOfOccupants" : (current_occupants + 1), "TenantSex" : tenant_sex}, "RoomNumber", tenant_room)
 
-            QMessageBox.information(self, "Success", "Tenant added successfully!")
+            QMessageBox.information(self, "Success", f"Tenant added successfully! \n Room {tenant_room} is now {current_occupants}/{maximum_capacity}")
             self.accept()
 
         except Exception as err:
